--- a/package.json
+++ b/package.json
@@ -94,7 +94,6 @@
     "safe-json-stringify": "^1.1.0",
     "shutdown": "^0.3.0",
     "stream-http": "^2.8.0",
-    "subcomandante": "^1.0.5",
     "superagent": "^3.8.2",
     "truthy": "0.0.1"
   },