@dryajov Some details for #188 

Well it was having a bad night, the 10s/command is now down to ~1.5s/command.  It's still slow.  

I'll be very interested in seeing what unix reports.  To get these results, you need to set the environment variable, in windows its `set DEBUG=ipfsd-ctl*`. Then `npm run test:node`.

```
  ipfsd-ctl:run ipfs.exe init -b 2048 +3s
  ipfsd-ctl:run ipfs.exe config API.HTTPHeaders.Access-Control-Allow-Origin ["*"] --json +2s
  ipfsd-ctl:run ipfs.exe config API.HTTPHeaders.Access-Control-Allow-Methods ["PUT","POST","GET"] --json +1s
  ipfsd-ctl:run ipfs.exe config Addresses.Swarm ["/ip4/127.0.0.1/tcp/0"] --json +1s
  ipfsd-ctl:run ipfs.exe config Addresses.API "/ip4/127.0.0.1/tcp/0" --json +1s
  ipfsd-ctl:run ipfs.exe config Addresses.Gateway "/ip4/127.0.0.1/tcp/0" --json +1s
          √ create node and init (7084ms)
  ipfsd-ctl:run ipfs.exe daemon +1s
          √ start node (3790ms)```