@richardschneider 2018-01-28
@diasdavid @dryajov `ipfsd-ctl` uses `subcommandte` to run the command.  It creates another NodeJS process that spawns the actual command.  Seem very inefficient to me.  `subcommandte` is fine when a long running child process needs to be killed when the parent the process is killed.  However, this is not our case, our child process are short lived and wait for them to close.

I'll try calling `child_process` directly and see what happens.