@richardschneider 2018-01-28
It's now ~350ms/command.

```
        spawn an initialized node
  ipfsd-ctl:run ipfs.exe init -b 2048 +2s
  ipfsd-ctl:run ipfs.exe config API.HTTPHeaders.Access-Control-Allow-Origin ["*"] --json +1s
  ipfsd-ctl:run ipfs.exe config API.HTTPHeaders.Access-Control-Allow-Methods ["PUT","POST","GET"] --json +303ms
  ipfsd-ctl:run ipfs.exe config Addresses.Swarm ["/ip4/127.0.0.1/tcp/0"] --json +286ms
  ipfsd-ctl:run ipfs.exe config Addresses.API "/ip4/127.0.0.1/tcp/0" --json +289ms
  ipfsd-ctl:run ipfs.exe config Addresses.Gateway "/ip4/127.0.0.1/tcp/0" --json +291ms
          √ create node and init (2584ms)
  ipfsd-ctl:run ipfs.exe daemon +288ms
```