@diasdavid 2018-01-28
@richardschneider historically, we had many issues with ipfs process spawning and killing, to the point that using `ipfs-desktop` would spawn dozens of IPFS nodes without being able to effectively shut them down.

In order to move away from subcomandante, we need to be extremely careful and test well that we don't end up in that world again.