@richardschneider 2018-01-28
@diasdavid "Test well", does that mean that all tests run on all platforms?