@richardschneider 2018-01-28
@diasdavid ran the `js-ipfs-api` tests in node and except for the 2 known pubsub test issues, everything passed.  I also checked for zombie processes and none were found.  During the test, at most only 4 js/go processes were active.

Then tried running the tests in the browser and as you can guess `webpack` failed with `heap error`.

I don't know about `ipfs-desktop`. How should I go about testing it?