@dryajov 2018-01-29
@richardschneider I wonder why such a performance bust? I'd imagine go being a bit lighter than node in general?