@richardschneider 2018-01-29
@dryajov I beginning to feel that killing a `node subprocess` is causing the issue, especially on Windows.

I'm fighting other fires right now, so will eventually get back to this.