@diasdavid 2018-02-16
> I don't know about ipfs-desktop. How should I go about testing it?

@hacdias can you give a hand here?