@richardschneider 2018-02-16
@diasdavid I've been waiting for `ipfsd-ctl` to settle down before continuing with this PR.

I'll rebase real soon and continue my investigations.