@richardschneider 2018-02-23
@diasdavid I need [graceful stop](https://github.com/ipfs/js-ipfsd-ctl/pull/205) before rebasing. What does another week matter.