@diasdavid 2018-03-14
@richardschneider time to rebase master onto this branch! :D