@diasdavid 2018-03-15
I've rebased master onto this branch, here are my results:

```
# master
npm test  159.45s user 27.56s system 65% cpu 4:45.73 total
# perf
npm test  134.43s user 21.23s system 63% cpu 4:06.89 total
```