@dryajov 2018-03-16
This seems to work as expected. Should we go with it (I'm 👍)?

@richardschneider if we merge this, lets not forget to remove subcomandante from the package.json deps