@diasdavid 2018-03-17
Please take into account my comment above -- https://github.com/ipfs/js-ipfsd-ctl/pull/190#issuecomment-361034791 -- simply trusting CI won't do it for this PR.