@dryajov 2018-03-20
@hacdias how can we test this with ipfs-desktop?