@travisperson 2018-04-05
**TL;DR;** - I think we need to look into this further to make sure we are handling things correctly and allowing gracefully shutdown across all platforms, while also being a friendly module to users consuming `ipfsd-ctl`.

---

I believe it is generally frowned upon to listen to the main process signals in a library. The issue is that when listening on `SIGINT` and `SIGTERM` the default node handlers are removed that cause the process to exit. Any module / application must implement its own handlers if it wishes to exit from the signals when the event loop is not empty if it consumes this module.

The `subcomandante` module gets around this as it does not listen on these signals from the consumed module, but from the forked child `subcom`, and polls for the parent `PID`.

It's a little dense and hard to parse, but the section on [Signal Events](https://nodejs.org/api/process.html#process_signal_events) provides some insight to how nodejs signals are handled.

I think it's also important to note this line from the process docs with regard to windows

> Sending SIGINT, SIGTERM, and SIGKILL cause the unconditional termination of the target process.

The standard `ChildProcess.prototype.kill` sends a `SIGTERM`, which cannot be handled on windows, `subcomandante` doesn't really handle this well either. It will turn a `SIGKILL` into a `SIGHUP` though.

Both nodejs and golang exit the process on a `SIGHUP`.

_Side note of Windows_
> `SIGHUP` is generated on Windows when the console window is closed [...] however, Node.js will be unconditionally terminated by Windows about 10 seconds later

The `go-ipfs` project gracefully handles the [`SIGHUP`](https://github.com/ipfs/go-ipfs/blob/master/cmd/ipfs/main.go#L408). However, `js-ipfs` [does not](https://github.com/ipfs/js-ipfs/blob/master/src/cli/commands/daemon.js#L54,L55) and instead will immediately terminate.

I don't have a good solution to this at the moment, but I do think we need to dig a bit further to see if we can come up with a better solution.

---

This thread (https://github.com/nodejs/node/issues/12378) on the node project provides more insight as well, and is worth a read through.

Signals in golang can also be read up on (https://golang.org/pkg/os/signal) here.