@dryajov 2018-04-05
I agree, we need to look into this further. I think that by removing the intermediate subcomandante process we fall back into the same issues subcomandante is trying to mitigate - namely forcefully killing ipfsd-ctl will leave a zombi ipfs process.

We should definitely handle the `SIGHUP` signal in js-ipfs, and it even works on windows - **amazing** 👅 !