@diasdavid 2018-06-04
I'm closing this PR given that the risk / speed tradeoff doesn't justify spending much more time into it right now.